refactor: js -> ts
