+++ src/main.ts
@@ -1,4 +1,5 @@
-import './assets/main.css'
+import "@/assets/scss/main.scss";
+import "@/assets/scss/buttons.scss";
 
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
