+++ src/settings/lang.ts
